+++ alfred-telemetry-solr/alfred-telemetry-solr6/src/main/java/eu/xenit/alfred/telemetry/solr/handler/MicrometerHandler.java
@@ -4,25 +4,30 @@ import eu.xenit.alfred.telemetry.solr.monitoring.binder.JvmMetrics;
 import eu.xenit.alfred.telemetry.solr.monitoring.binder.ProcessMetrics;
 import eu.xenit.alfred.telemetry.solr.monitoring.binder.SolrMetrics;
 import eu.xenit.alfred.telemetry.solr.monitoring.binder.SystemMetrics;
-import eu.xenit.alfred.telemetry.solr.monitoring.registry.RegistryRegistraar;
+import eu.xenit.alfred.telemetry.solr.monitoring.registry.RegistryRegistrar;
 import eu.xenit.alfred.telemetry.solr.util.PrometheusRegistryUtil;
 import eu.xenit.alfred.telemetry.solr.util.Util;
 import io.micrometer.core.instrument.MeterRegistry;
-import java.io.IOException;
-import javax.management.MBeanServer;
+import io.micrometer.core.instrument.Tags;
+import io.micrometer.core.instrument.binder.jetty.JettyStatisticsMetrics;
 import org.alfresco.solr.AlfrescoCoreAdminHandler;
 import org.apache.solr.handler.RequestHandlerBase;
 import org.apache.solr.request.SolrQueryRequest;
 import org.apache.solr.response.SolrQueryResponse;
 import org.apache.solr.util.JmxUtil;
+import org.eclipse.jetty.server.Server;
+import org.eclipse.jetty.server.handler.StatisticsHandler;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import javax.management.MBeanServer;
+import java.io.IOException;
+
 public class MicrometerHandler extends RequestHandlerBase {
 
-    static RegistryRegistraar registraar = new RegistryRegistraar();
-    static MeterRegistry registry = registraar.getGlobalMeterRegistry();
+    static MeterRegistry registry = RegistryRegistrar.getInstance().getGlobalMeterRegistry();
     static SolrMetrics solrMetrics = null;
+    static JettyStatisticsMetrics jettyMetrics = null;
 
     static {
         if( Util.isEnabled("ALFRED_TELEMETRY_JVM_ENABLED"))
@@ -42,11 +47,26 @@ public class MicrometerHandler extends RequestHandlerBase {
                 .getMultiCoreHandler();
         MBeanServer mbeanServer = JmxUtil.findFirstMBeanServer();
 
+        if(jettyMetrics == null && Util.isEnabled("METRICS_JETTY_ENABLED")) {
+            Server server =  (Server)req.getHttpSolrCall().getReq().getAttribute("org.eclipse.jetty.server.Server");
+            if(server==null) {
+                logger.error("There is no jetty server");
+            } else {
+                StatisticsHandler statisticsHandler = (StatisticsHandler) server.getChildHandlerByClass(StatisticsHandler.class);
+                if(statisticsHandler==null) {
+                    logger.error("There is no StatisticsHandler");
+                } else{
+                    jettyMetrics = new JettyStatisticsMetrics(statisticsHandler, Tags.empty());
+                    jettyMetrics.bindTo(registry);
+                }
+            }
+        }
+
         if (solrMetrics == null && Util.isEnabled("METRICS_SOLR_ENABLED")) {
             solrMetrics = new SolrMetrics(coreAdminHandler, mbeanServer);
             solrMetrics.bindTo(registry);
         }
-        writeTextToResponse(PrometheusRegistryUtil.extractPrometheusScrapeData(registraar.getPrometheusMeterRegistry()),
+        writeTextToResponse(PrometheusRegistryUtil.extractPrometheusScrapeData(RegistryRegistrar.getInstance().getPrometheusMeterRegistry()),
                 rsp);
     }